refactor: extract unified ParseVerdict from ParseReviewResult and ParseQualityResult
The two functions were structurally identical, differing only in the
keyword ("REVIEW" vs "QUALITY"). Introduce ParseVerdict(stdout, keyword)
in verdict.go and make both existing functions thin wrappers, preserving
backwards compatibility. Migrate all 11 verdict-specific tests to
verdict_test.go.

Closes #390
