+++ internal/agent/guardrails_test.go
@@ -8,30 +8,6 @@ import (
 	"testing"
 )
 
-func TestParseReviewResult_Approved(t *testing.T) {
-	stdout := "some output\nREVIEW_RESULT=APPROVED\nmore output"
-	got := ParseReviewResult(stdout)
-	if got != "APPROVED" {
-		t.Errorf("ParseReviewResult() = %q, want %q", got, "APPROVED")
-	}
-}
-
-func TestParseReviewResult_ChangesRequested(t *testing.T) {
-	stdout := "output\nREVIEW_RESULT=CHANGES_REQUESTED\n"
-	got := ParseReviewResult(stdout)
-	if got != "CHANGES_REQUESTED" {
-		t.Errorf("ParseReviewResult() = %q, want %q", got, "CHANGES_REQUESTED")
-	}
-}
-
-func TestParseReviewResult_NotFound(t *testing.T) {
-	stdout := "just some random output\nno result here"
-	got := ParseReviewResult(stdout)
-	if got != "" {
-		t.Errorf("ParseReviewResult() = %q, want empty", got)
-	}
-}
-
 func TestFindPR_ReturnsPRNumber(t *testing.T) {
 	stubGuardRunner(t, func(name string, args ...string) ([]byte, error) {
 		return []byte(`{"number": 42}`), nil
@@ -284,46 +260,6 @@ func TestClosePR_ReturnsError(t *testing.T) {
 	}
 	if !strings.Contains(err.Error(), "closing PR #10") {
 		t.Errorf("error = %v, want 'closing PR #10'", err)
-	}
-}
-
-func TestParseReviewResult_FirstMatchWins(t *testing.T) {
-	stdout := "REVIEW_RESULT=APPROVED\nREVIEW_RESULT=CHANGES_REQUESTED\n"
-	got := ParseReviewResult(stdout)
-	if got != "APPROVED" {
-		t.Errorf("ParseReviewResult() = %q, want %q (first match should win)", got, "APPROVED")
-	}
-}
-
-func TestParseReviewResult_WhitespaceHandling(t *testing.T) {
-	stdout := "  REVIEW_RESULT=APPROVED  \n"
-	got := ParseReviewResult(stdout)
-	if got != "APPROVED" {
-		t.Errorf("ParseReviewResult() = %q, want %q", got, "APPROVED")
-	}
-}
-
-func TestParseReviewResult_ColonFormat(t *testing.T) {
-	stdout := "REVIEW RESULT: APPROVED\n"
-	got := ParseReviewResult(stdout)
-	if got != "APPROVED" {
-		t.Errorf("ParseReviewResult() = %q, want %q", got, "APPROVED")
-	}
-}
-
-func TestParseReviewResult_ColonFormatChanges(t *testing.T) {
-	stdout := "REVIEW RESULT: CHANGES_REQUESTED\n"
-	got := ParseReviewResult(stdout)
-	if got != "CHANGES_REQUESTED" {
-		t.Errorf("ParseReviewResult() = %q, want %q", got, "CHANGES_REQUESTED")
-	}
-}
-
-func TestParseReviewResult_CaseInsensitive(t *testing.T) {
-	stdout := "Review Result: Approved\n"
-	got := ParseReviewResult(stdout)
-	if got != "APPROVED" {
-		t.Errorf("ParseReviewResult() = %q, want %q", got, "APPROVED")
 	}
 }
